wip: put each cmd into their own module
